pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opendms"
version = "0.1.0"
description = "Open Document Management System with VeriDocs DID/VC integration"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "PyJWT>=2.9.0",
    "python-multipart>=0.0.12",
    "boto3>=1.35.0",
]

[project.optional-dependencies]
azure = ["azure-storage-blob>=12.0.0"]
search = ["elasticsearch[async]>=8.0.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24.0", "ruff>=0.8.0"]

[tool.setuptools.packages.find]
where = ["src"]
